Disable ADHDinos for now
The API endpoint we relied on was taken down in favor of server-side
page generation, and the "Latest Comics" link now takes you directly to
Instagram.

ADHDinos will be back when we switch the scraping script to reddit. In
the meantime the existing index will remain available, hopefully the IG
CDN links don't go stale. :meowSweats: